% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_delete_button.R
\name{addRowAt}
\alias{addRowAt}
\title{Adds a row at a specified index}
\usage{
addRowAt(df, row, i)
}
\arguments{
\item{df}{a data frame}

\item{row}{a row with the same columns as \code{df}}

\item{i}{the index we want to add row at.}
}
\value{
the data frame with \code{row} added to \code{df} at index \code{i}
}
\description{
Adds a row at a specified index
}
