% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_package_data.R
\docType{data}
\name{days_of_the_week}
\alias{days_of_the_week}
\title{Data: Days of the Week}
\format{
A factor vector.
}
\usage{
days_of_the_week
}
\description{
Vector of days of the week for date calculations.
}
\keyword{datasets}
