% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinnR.R
\docType{data}
\name{dinn}
\alias{dinn}
\title{Neccessary Ingredients for DinnR}
\format{
A data frame with a varying count of rows and 7 variables
}
\source{
\url{https://docs.google.com/spreadsheets/d/1qVV-QL7l0WXnf0kbK29qOvv_y_6lIgnm_xaWG8QxYoE/edit?usp=sharing}
}
\usage{
dinn
}
\description{
A data set that is read in from Google Sheets.
}
\keyword{datasets}
