% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinnR-package.R
\docType{package}
\name{dinnR-package}
\alias{dinnR}
\alias{dinnR-package}
\title{dinnR: What's for Dinner?}
\description{
Summarize what ingredients needed for the week to cook the
    food!
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KoderKow/dinnR}
  \item Report bugs at \url{https://github.com/KoderKow/dinnR/issues}
}

}
\author{
\strong{Maintainer}: Kyle Harris \email{koderkow@gmail.com}

}
\keyword{internal}
