% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_planning_dates}
\alias{get_planning_dates}
\title{Find the date for the wanted next week day}
\usage{
get_planning_dates(starting_date = NULL, next_monday = TRUE)
}
\arguments{
\item{starting_date}{A character. Date you are starting from.}

\item{next_monday}{A logical. Should the vector start with the following monday?}
}
\description{
Thank you `@aftonsteps` for the simplified code!
}
\references{
https://gist.github.com/aftonsteps/1aad7b0220e9ce8e895020a105ed3386
}
