% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_theme_538.R
\name{gt_dinnR}
\alias{gt_dinnR}
\title{Custom {gt} Styles}
\usage{
gt_dinnR(.data)
}
\arguments{
\item{.data}{a gt object.}
}
\value{
a gt object
}
\description{
Some extra styling for the tables used in this app.
}
