% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_theme_538.R
\name{gt_theme_538}
\alias{gt_theme_538}
\title{538 {gt} Theme}
\usage{
gt_theme_538(data, ...)
}
\arguments{
\item{data}{gt object.}

\item{...}{Additional params to pass to [gt::tab_options()]}
}
\value{
a gt object
}
\description{
Amazing theme made by [Tom Mock](https://themockup.blog/posts/2020-09-26-functions-and-themes-for-gt-tables/)
}
