% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_package_data.R
\docType{data}
\name{last_updated}
\alias{last_updated}
\title{Data: Last Updated}
\format{
A date object.
}
\usage{
last_updated
}
\description{
Date to show when the data was last updated.
}
\keyword{datasets}
