% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{next_week_day}
\alias{next_week_day}
\title{Find the date for the wanted next week day}
\usage{
next_week_day(starting_date, week_day)
}
\arguments{
\item{starting_date}{A character. Date you are starting from.}

\item{week_day}{A numeric. The next week day you want the date for.}
}
\description{
Find the date for the wanted next week day
}
\references{
https://stackoverflow.com/questions/32434549/how-to-find-next-particular-day
}
