% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{random_recipe}
\alias{random_recipe}
\title{Retrieve Random Recipes}
\usage{
random_recipe(r, n)
}
\arguments{
\item{r}{A reactive list.}

\item{n}{A numeric. Number of random recipes to recieve.}
}
\description{
Retrieve Random Recipes
}
